move signin controller function into common folder Relates #20
