+++ server/controller/common/index.js
@@ -1,0 +1,3 @@
+const signin = require('./signin');
+
+module.exports = { signin };
