--- server/controller/user/index.js
@@ -1,3 +1,0 @@
-const signin = require('./signin');
-
-module.exports = { signin };
